@@ -1,12 +1,12 @@
 [project]
 name = "gamelib"
-version = "0.1.0"
+version = "0.1.1"
 description = ""
 authors = [
     {name = "welc0186",email = "49768046+welc0186@users.noreply.github.com"}
 ]
 readme = "README.md"
-requires-python = ">=3.10"
+requires-python = ">=3.11"
 dependencies = [
     "pygame (>=2.6.1,<3.0.0)",
     "esper (>=3.4,<4.0)"
